refactor: replace SearchForm component
